requirements.txt
adafruit-ampy==1.1.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
debugpy==1.8.8
Django==5.1.2
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
paho-mqtt==2.1.0
parse_pip_search==0.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
redis==5.2.0
requests==2.32.3
rich==13.9.2
service-identity==24.2.0
setuptools==75.5.0
simple-websocket==1.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tk==0.1.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
zope.interface==7.1.1
